Moved the main function to a main.go file
